Remove old functional tests from `npm test` script
The functional tests are outdated and should be reviewed before adding
them back in.

NOTE: istanbul cover tool is having issues with npm which causes an
error to occur when writing the coverage report. If a coverage report is
not being created but needed, try running the coverage manually, e.g.
`./node_modules/.bin/istanbul cover -- ./test/unit/**
./test/integration/**`